% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMasterR.R
\name{createMasterR}
\alias{createMasterR}
\title{createMasterR}
\usage{
createMasterR(drive, batch, EMG = FALSE, tagIDs)
}
\arguments{
\item{drive}{The single-letter name of the drive where the database is located. Usually "G" or "E" depending on which computer you are working from.}

\item{batch}{The name or names of the batch folders where the receiver text files are stored.}

\item{EMG}{Defaults to FALSE, if TRUE, the data will be formatted for EMG files, not CART files.}

\item{tagIDs}{Unless you specified EMG = TRUE, a single C.Code.ID or a vector of C.Code.IDs with the TagIDs you wish to examine. Best to set this as the C.Code.ID column from your fish data workbook dataframe. If you set EMG=T, then use the EMG.ID column, not the C.Code.ID.}
}
\description{
This function allows you to compile data from batches containing radio receiver text files.
}
\examples{
createMasterR(drive="G",batch="CLRADIO",EMG=F,tagIDs=fish$C.Code.ID)
}
\keyword{telemetry}
